+++ packages/plugin-dva/src/index.ts
@@ -17,9 +17,13 @@ export default (api: IApi) => {
     return join(api.paths.absSrcPath!, getModelDir());
   }
 
-  function hasDvaDependency() {
+  function getDvaDependency() {
     const { dependencies, devDependencies } = api.pkg;
-    return dependencies?.dva || devDependencies?.dva;
+    return (
+      dependencies?.dva ||
+      devDependencies?.dva ||
+      require('../package').dependencies.dva
+    );
   }
 
   // 配置
@@ -28,10 +32,11 @@ export default (api: IApi) => {
     config: {
       schema(joi) {
         return joi.object({
-          immer: joi.boolean(),
-          hmr: joi.boolean(),
-          skipModelValidate: joi.boolean(),
+          disableModelsReExport: joi.boolean(),
           extraModels: joi.array().items(joi.string()),
+          hmr: joi.boolean(),
+          immer: joi.alternatives(joi.boolean(), joi.object()),
+          skipModelValidate: joi.boolean(),
         });
       },
     },
@@ -46,15 +51,18 @@ export default (api: IApi) => {
     return lodash.uniq([
       ...getModels({
         base: srcModelsPath,
+        cwd: api.cwd,
         ...baseOpts,
       }),
       ...getModels({
         base: api.paths.absPagesPath!,
+        cwd: api.cwd,
         pattern: `**/${getModelDir()}/**/*.{ts,tsx,js,jsx}`,
         ...baseOpts,
       }),
       ...getModels({
         base: api.paths.absPagesPath!,
+        cwd: api.cwd,
         pattern: `**/model.{ts,tsx,js,jsx}`,
         ...baseOpts,
       }),
@@ -66,6 +74,13 @@ export default (api: IApi) => {
   // 初始检测一遍
   api.onStart(() => {
     hasModels = getAllModels().length > 0;
+  });
+
+  api.addDepInfo(() => {
+    return {
+      name: 'dva',
+      range: getDvaDependency(),
+    };
   });
 
   // 生成临时文件
@@ -87,17 +102,24 @@ export default (api: IApi) => {
         content: Mustache.render(dvaTpl, {
           ExtendDvaConfig: '',
           EnhanceApp: '',
-          RegisterPlugins: [
-            // immutable state
-            api.config.dva?.immer &&
-              `app.use(require('${winPath(require.resolve('dva-immer'))}')());`,
-          ]
-            .filter(Boolean)
-            .join('\n'),
+          dvaImmer: api.config.dva?.immer,
+          dvaImmerPath: winPath(require.resolve('dva-immer')),
+          dvaImmerES5:
+            lodash.isPlainObject(api.config.dva?.immer) &&
+            api.config.dva?.immer.enableES5,
+          RegisterModelImports: models
+            .map((path, index) => {
+              return `import Model${lodash.upperFirst(
+                lodash.camelCase(basename(path, extname(path))),
+              )}${index} from '${path}';`;
+            })
+            .join('\r\n'),
           RegisterModels: models
-            .map(path => {
-              // 注册用户 model
-              return `app.model({ namespace: '${basename(path, extname(path))}', ...(require('${path}').default) });`.trim();
+            .map((path, index) => {
+              // prettier-ignore
+              return `
+app.model({ namespace: '${basename(path, extname(path))}', ...Model${lodash.upperFirst(lodash.camelCase(basename(path, extname(path))))}${index} });
+          `.trim();
             })
             .join('\r\n'),
           // use esm version
@@ -130,7 +152,6 @@ export default (api: IApi) => {
         ? ['connect', 'useDispatch', 'useStore', 'useSelector']
         : ['connect'];
 
-      logger.debug(`dva lib path: ${dvaLibPath}`);
       logger.debug(`dva version: ${dvaVersion}`);
       logger.debug(`exported methods:`);
       logger.debug(exportMethods);
@@ -138,7 +159,6 @@ export default (api: IApi) => {
       api.writeTmpFile({
         path: 'plugin-dva/exports.ts',
         content: Mustache.render(exportsTpl, {
-          dvaLibPath,
           exportMethods: exportMethods.join(', '),
         }),
       });
@@ -149,14 +169,16 @@ export default (api: IApi) => {
       api.writeTmpFile({
         path: 'plugin-dva/connect.ts',
         content: Mustache.render(connectTpl, {
-          dvaHeadExport: models
-            .map(path => {
-              // prettier-ignore
-              return `export * from '${winPath(dirname(path) + "/" + basename(path, extname(path)))}';`;
-            })
-            .join('\r\n'),
+          dvaHeadExport: api.config.dva?.disableModelsReExport
+            ? ``
+            : models
+                .map((path) => {
+                  // prettier-ignore
+                  return `export * from '${winPath(dirname(path) + "/" + basename(path, extname(path)))}';`;
+                })
+                .join('\r\n'),
           dvaLoadingModels: models
-            .map(path => {
+            .map((path) => {
               // prettier-ignore
               return `    ${basename(path, extname(path))
                 } ?: boolean;`;
@@ -179,7 +201,7 @@ export default (api: IApi) => {
   ]);
 
   // Babel Plugin for HMR
-  api.modifyBabelOpts(babelOpts => {
+  api.modifyBabelOpts((babelOpts) => {
     const hmr = api.config.dva?.hmr;
     if (hmr) {
       const hmrOpts = lodash.isPlainObject(hmr) ? hmr : {};
@@ -221,7 +243,7 @@ export default (api: IApi) => {
         console.log();
         console.log(utils.chalk.bold('  Models in your project:'));
         console.log();
-        models.forEach(model => {
+        models.forEach((model) => {
           console.log(`    - ${relative(api.cwd, model)}`);
         });
         console.log();