+++ scripts/release.js
@@ -105,10 +105,14 @@ async function release() {
         // '--no-push',
         '--message',
         'chore(release): Publish',
-        '--conventional-commits',
+        args.selectVersion ? '' : '--conventional-commits',
       ]
+        .filter(Boolean)
         .concat(conventionalGraduate)
         .concat(conventionalPrerelease),
+      {
+        shell: false,
+      },
     );
   }