+++ packages/plugin-dva/CHANGELOG.md
@@ -2,6 +2,58 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.11.0](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.10.0...@umijs/plugin-dva@0.11.0) (2021-03-26)
+
+### Features
+
+- improve dva log ([#562](https://github.com/umijs/plugins/issues/562)) ([4d84817](https://github.com/umijs/plugins/commit/4d84817a041fd259afb6a8c7f6abfac16e7925e5))
+
+# [0.10.0](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.9.1...@umijs/plugin-dva@0.10.0) (2021-01-13)
+
+### Features
+
+- plugin dva support config immer.enableES5 ([#505](https://github.com/umijs/plugins/issues/505)) ([969d9d8](https://github.com/umijs/plugins/commit/969d9d8b3ba4f0e9cdfc9c6854adfed0458fcb6d))
+
+## [0.9.1](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.9.0...@umijs/plugin-dva@0.9.1) (2020-09-24)
+
+**Note:** Version bump only for package @umijs/plugin-dva
+
+# [0.9.0](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.8.2...@umijs/plugin-dva@0.9.0) (2020-09-15)
+
+### Features
+
+- **dva:** support config disableModelsReExport, don't reexport model in connect.ts ([#366](https://github.com/umijs/plugins/issues/366)) ([0f40f1d](https://github.com/umijs/plugins/commit/0f40f1d038e04ebc959a49d62a42b78177f14995))
+
+## [0.8.2](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.8.1...@umijs/plugin-dva@0.8.2) (2020-08-25)
+
+### Bug Fixes
+
+- **dva:** upperCamelCase model's default specifier ([#342](https://github.com/umijs/plugins/issues/342)) ([ef3fc39](https://github.com/umijs/plugins/commit/ef3fc39209abf445b5670cc829099b6aad2d2de1))
+
+## [0.8.1](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.8.0...@umijs/plugin-dva@0.8.1) (2020-08-25)
+
+### Bug Fixes
+
+- dva model register duplicate ([#340](https://github.com/umijs/plugins/issues/340)) ([4845db8](https://github.com/umijs/plugins/commit/4845db80b0521c13396f05f064b4742859fa26f2))
+
+# [0.8.0](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.7.4...@umijs/plugin-dva@0.8.0) (2020-08-25)
+
+### Features
+
+- enhancement for mfsu ([#326](https://github.com/umijs/plugins/issues/326)) ([655c0da](https://github.com/umijs/plugins/commit/655c0da475748a0671dd3a5de8ab079dbe1bed5a))
+
+## [0.7.4](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.7.3...@umijs/plugin-dva@0.7.4) (2020-07-01)
+
+### Bug Fixes
+
+- **plugin-dva:** add RequiredConnectProps ([#285](https://github.com/umijs/plugins/issues/285)) ([fd2ba9c](https://github.com/umijs/plugins/commit/fd2ba9cb11854adc6c4d055454fdcbcbc41f2c01))
+
+## [0.7.3](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.7.2...@umijs/plugin-dva@0.7.3) (2020-06-30)
+
+### Bug Fixes
+
+- **plugin-dva:** add query props ([#278](https://github.com/umijs/plugins/issues/278)) ([d2d9cb6](https://github.com/umijs/plugins/commit/d2d9cb6cc7a05791b8e1d374fedb1f79d5dae8df))
+
 ## [0.7.2](https://github.com/umijs/plugins/compare/@umijs/plugin-dva@0.7.1...@umijs/plugin-dva@0.7.2) (2020-05-21)
 
 ### Bug Fixes