+++ packages/plugin-esbuild/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/umijs/plugins/compare/@umijs/plugin-esbuild@1.1.0...@umijs/plugin-esbuild@1.1.1) (2021-03-26)
+
+### Bug Fixes
+
+- **esbuild:** minify exposes target config for es5 projects ([#571](https://github.com/umijs/plugins/issues/571)) ([cb93b3a](https://github.com/umijs/plugins/commit/cb93b3a969d9f94c808eda8e02828f347cc2eca0))
+
+## [1.0.3](https://github.com/umijs/plugins/compare/@umijs/plugin-esbuild@1.0.2...@umijs/plugin-esbuild@1.0.3) (2020-12-15)
+
+### Bug Fixes
+
+- esbuild platform undefined ([#477](https://github.com/umijs/plugins/issues/477)) ([1fd2887](https://github.com/umijs/plugins/commit/1fd2887ee849a2ff079f265645f49d5920cc3fda))
+
+## [1.0.2](https://github.com/umijs/plugins/compare/@umijs/plugin-esbuild@1.0.1...@umijs/plugin-esbuild@1.0.2) (2020-12-15)
+
+### Bug Fixes
+
+- esbuild ssr ([#309](https://github.com/umijs/plugins/issues/309)) ([68ca76f](https://github.com/umijs/plugins/commit/68ca76fd22e23a67e9928d2addb9209c9955ef42))
+
+## [1.0.1](https://github.com/umijs/plugins/compare/@umijs/plugin-esbuild@1.0.0...@umijs/plugin-esbuild@1.0.1) (2020-08-21)
+
+### Bug Fixes
+
+- esbuild targets ([#336](https://github.com/umijs/plugins/issues/336)) ([34c3a1d](https://github.com/umijs/plugins/commit/34c3a1dcc0bcb81cb13cfe36a306cfc1d9ecedc7))
+
 # [1.0.0](https://github.com/umijs/plugins/compare/@umijs/plugin-esbuild@1.0.0-beta.3...@umijs/plugin-esbuild@1.0.0) (2020-05-19)
 
 **Note:** Version bump only for package @umijs/plugin-esbuild