+++ packages/preset-react/CHANGELOG.md
@@ -2,6 +2,239 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.3](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.2...@umijs/preset-react@1.8.3) (2021-03-26)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.2](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.1...@umijs/preset-react@1.8.2) (2021-03-19)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.1](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.0...@umijs/preset-react@1.8.1) (2021-03-15)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+# [1.8.0](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.13...@umijs/preset-react@1.8.0) (2021-03-04)
+
+### Features
+
+- **lodash:** on demand loaded ([#535](https://github.com/umijs/plugins/issues/535)) ([0dd2998](https://github.com/umijs/plugins/commit/0dd29985bf4a9669ea43606c876c8dec3f35178a))
+
+### Reverts
+
+- Revert "feat(lodash): on demand loaded (#535)" (#539) ([d931276](https://github.com/umijs/plugins/commit/d93127667e2c696b410312f59f883f4e887eb6ff)), closes [#535](https://github.com/umijs/plugins/issues/535) [#539](https://github.com/umijs/plugins/issues/539)
+
+## [1.7.13](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.12...@umijs/preset-react@1.7.13) (2021-01-19)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.12](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.11...@umijs/preset-react@1.7.12) (2021-01-13)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.11](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.10...@umijs/preset-react@1.7.11) (2021-01-13)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.10](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.9...@umijs/preset-react@1.7.10) (2021-01-11)
+
+### Bug Fixes
+
+- remove babel-plugin-react-require when using automatic runtime ([#497](https://github.com/umijs/plugins/issues/497)) ([5d9ba02](https://github.com/umijs/plugins/commit/5d9ba02aab752a12c7505eb60f65843f3d6df84c))
+- 修复在 React 17 下当不使用 ts 的情况下，启动项目仍然会提示更新 tsconfig.json 中 jsx 字段的问题 ([#482](https://github.com/umijs/plugins/issues/482)) ([84f9a91](https://github.com/umijs/plugins/commit/84f9a919efb9b28e854912429039b8d1e1a5be7a))
+
+## [1.7.9](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.8...@umijs/preset-react@1.7.9) (2020-12-24)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.8](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.7...@umijs/preset-react@1.7.8) (2020-12-10)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.7](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.6...@umijs/preset-react@1.7.7) (2020-12-08)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.6](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.5...@umijs/preset-react@1.7.6) (2020-12-03)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.5](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.4...@umijs/preset-react@1.7.5) (2020-12-03)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.4](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.3...@umijs/preset-react@1.7.4) (2020-12-02)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.3](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.2...@umijs/preset-react@1.7.3) (2020-12-02)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.2](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.1...@umijs/preset-react@1.7.2) (2020-12-02)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.7.1](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.7.0...@umijs/preset-react@1.7.1) (2020-12-01)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+# [1.7.0](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.19...@umijs/preset-react@1.7.0) (2020-11-19)
+
+### Bug Fixes
+
+- react undefined ([#447](https://github.com/umijs/plugins/issues/447)) ([fe71be1](https://github.com/umijs/plugins/commit/fe71be1c1d08c6297bcf5cc1b3764633fe75bee0))
+
+### Features
+
+- support React 17 new JSX syntax ([#445](https://github.com/umijs/plugins/issues/445)) ([7cdb18d](https://github.com/umijs/plugins/commit/7cdb18d81bd802416bc0778f80fe2fa2976d54b0))
+
+## [1.6.19](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.18...@umijs/preset-react@1.6.19) (2020-11-06)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.18](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.17...@umijs/preset-react@1.6.18) (2020-11-06)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.17](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.16...@umijs/preset-react@1.6.17) (2020-11-04)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.16](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.15...@umijs/preset-react@1.6.16) (2020-10-27)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.15](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.14...@umijs/preset-react@1.6.15) (2020-10-26)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.14](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.13...@umijs/preset-react@1.6.14) (2020-10-22)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.13](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.12...@umijs/preset-react@1.6.13) (2020-10-18)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.12](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.11...@umijs/preset-react@1.6.12) (2020-10-16)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.11](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.10...@umijs/preset-react@1.6.11) (2020-10-16)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.10](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.9...@umijs/preset-react@1.6.10) (2020-10-12)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.9](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.8...@umijs/preset-react@1.6.9) (2020-10-09)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.8](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.7...@umijs/preset-react@1.6.8) (2020-09-24)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.7](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.6...@umijs/preset-react@1.6.7) (2020-09-21)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.6](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.5...@umijs/preset-react@1.6.6) (2020-09-17)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.5](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.4...@umijs/preset-react@1.6.5) (2020-09-15)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.4](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.3...@umijs/preset-react@1.6.4) (2020-08-27)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.3](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.2...@umijs/preset-react@1.6.3) (2020-08-26)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.2](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.1...@umijs/preset-react@1.6.2) (2020-08-25)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.6.1](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.6.0...@umijs/preset-react@1.6.1) (2020-08-25)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+# [1.6.0](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.22...@umijs/preset-react@1.6.0) (2020-08-25)
+
+### Features
+
+- enhancement for mfsu ([#326](https://github.com/umijs/plugins/issues/326)) ([655c0da](https://github.com/umijs/plugins/commit/655c0da475748a0671dd3a5de8ab079dbe1bed5a))
+
+## [1.5.22](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.21...@umijs/preset-react@1.5.22) (2020-08-19)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.21](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.20...@umijs/preset-react@1.5.21) (2020-08-12)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.20](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.19...@umijs/preset-react@1.5.20) (2020-08-12)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.19](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.18...@umijs/preset-react@1.5.19) (2020-07-20)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.18](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.17...@umijs/preset-react@1.5.18) (2020-07-20)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.17](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.16...@umijs/preset-react@1.5.17) (2020-07-09)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.16](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.15...@umijs/preset-react@1.5.16) (2020-07-07)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.15](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.14...@umijs/preset-react@1.5.15) (2020-07-03)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.14](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.13...@umijs/preset-react@1.5.14) (2020-07-02)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.13](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.12...@umijs/preset-react@1.5.13) (2020-07-01)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.12](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.11...@umijs/preset-react@1.5.12) (2020-06-30)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.11](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.10...@umijs/preset-react@1.5.11) (2020-06-28)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.10](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.9...@umijs/preset-react@1.5.10) (2020-06-22)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.9](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.8...@umijs/preset-react@1.5.9) (2020-06-20)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.8](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.7...@umijs/preset-react@1.5.8) (2020-06-15)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.5.7](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.6...@umijs/preset-react@1.5.7) (2020-06-14)
+
+**Note:** Version bump only for package @umijs/preset-react
+
 ## [1.5.6](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.5.5...@umijs/preset-react@1.5.6) (2020-06-03)
 
 **Note:** Version bump only for package @umijs/preset-react